requirements.txt
flask
pyannote-audio
sentencepiece